OOP & Docker. Begining
