+++ sender/requirements.txt
@@ -1,0 +1,6 @@
+bs4==0.0.1
+requests==2.22.0
+selenium==3.141.0
+pytesseract==0.2.9
+lxml==4.4.1
+PyYAML==5.1.2
